% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-phylo.R
\name{alml_2_phylo}
\alias{alml_2_phylo}
\title{Transform a alml result to phylo}
\usage{
alml_2_phylo(the_result)
}
\arguments{
\item{the_result}{one result which is read in from alml file}
}
\description{
Transform a alml result to phylo
}
