% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Bin2Dec}
\alias{Bin2Dec}
\title{Transform binary sequence to decimal sequence}
\usage{
Bin2Dec(seq)
}
\arguments{
\item{seq}{a binary sequence of character class}
}
\value{
a decimal number of numeric class
}
\description{
Transform binary sequence to decimal sequence
}
\examples{
Bin2Dec("10")
Bin2Dec("1000")
}
