% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Bin_get_parent}
\alias{Bin_get_parent}
\title{Get the binary sequence of parent}
\usage{
Bin_get_parent(x)
}
\arguments{
\item{seq}{a binary sequence of the daughter}
}
\value{
a binary sequence of the mother/parent
}
\description{
We can use the binary index to find the index of its parent.
}
