% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Dec2Bin}
\alias{Dec2Bin}
\title{Transform a decimal number to a binary sequence of character}
\usage{
Dec2Bin(x)
}
\arguments{
\item{x}{a decimal number}
}
\value{
a binary sequence of character
}
\description{
Transform a decimal number to a binary sequence of character
}
