% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELTA.R
\name{DELTA}
\alias{DELTA}
\title{Calculate DELTA score}
\usage{
DELTA(treeS, treeT, cost, outfile = "./", method = "g", max_target = 1,
  test = 0, all = "F", prune = 1, DELTA.address)
}
\arguments{
\item{treeS:}{TreeS file path.(required)}

\item{treeT:}{TreeT file path.(required)}

\item{cost:}{Cost file path. the cost file contains the score for different types of leaves.(required)}

\item{method:}{l or g. g for global alignment; l for local alignment. default: g.}

\item{max_target:}{target num for l, local alignment. default: 1}

\item{test:}{testNum to calculate p-value. If testNum <=2, do not output p-value. default 0}

\item{outfile:}{output file path. default: TreeS file path + l or g, based on -method}

\item{all:}{T or F. output as much information as possible. default F;}

\item{prune:}{pruneScore is the punish for pruning one leaf. default 1}
}
\description{
Calculate DELTA score
}
