% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawScoreMatrix.R
\name{DrawScoreMatrix2}
\alias{DrawScoreMatrix2}
\title{Draw score matrix}
\usage{
DrawScoreMatrix2(alml_list, n = 48)
}
\arguments{
\item{alml_list}{result read from readal}

\item{n}{rainbow max num for scale. heatmap color style.}
}
\description{
Draw score matrix
}
