% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find-addNode.R
\name{Find.addNode}
\alias{Find.addNode}
\title{Find the missing nodes, including leaves and internal nodes.}
\usage{
Find.addNode(allLin)
}
\description{
Find the missing nodes, including leaves and internal nodes.
}
