% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find-missed-sister.R
\name{Find.missed.sister}
\alias{Find.missed.sister}
\title{Find missed sister in all lineaged tree.}
\usage{
Find.missed.sister(allLin)
}
\arguments{
\item{allLin}{the lineage vector of this tree, including internal nodes and leaves.}
}
\description{
Find missed sister in all lineaged tree.
}
