% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find-root-from-tips.R
\name{Find.root.from.tips}
\alias{Find.root.from.tips}
\title{Find root from tips}
\usage{
Find.root.from.tips(LeafLin)
}
\arguments{
\item{LeafLin}{Lineage of tips}
}
\description{
Find root from tips
}
