% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find-tips.R
\name{Find.tips}
\alias{Find.tips}
\title{Find tips from all lineage of complete tree.}
\usage{
Find.tips(allLin)
}
\arguments{
\item{allLin}{all lineage of tree}
}
\value{
tips lineage
}
\description{
Find tips from all lineage of complete tree.
}
