% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find-tree-from-tips.R
\name{Find.tree.from.tips}
\alias{Find.tree.from.tips}
\title{Rebuild the complete tree from tips (leaves) lineage.}
\usage{
Find.tree.from.tips(allTips)
}
\arguments{
\item{allTips}{all lineage of tips(leaves).}
}
\value{
all lineage of a complete tree.
}
\description{
Rebuild the complete tree from tips (leaves) lineage.
}
