% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find-unpaired-tips.R
\name{Find.unpaired.tip}
\alias{Find.unpaired.tip}
\title{Find unpaired tips of an incompleted tree all lineage.}
\usage{
Find.unpaired.tip(allLin)
}
\arguments{
\item{allLin}{all lineage}
}
\value{
unpaired tips
}
\description{
Find unpaired tips of an incompleted tree all lineage.
}
