% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_EPIC.R
\name{geom_EPIC}
\alias{geom_EPIC}
\title{geom_EPIC}
\usage{
geom_EPIC(expr_file, expr_size = 1.5, expr_alpha = 1, mc.cores = 5,
  color_gradientn = c("black", "grey", "darkblue", "blue", "skyblue",
  "purple", "red", "orange", "green", "yellow"), values = c(0, 0.01, 0.02,
  0.04, 0.06, 0.08, 0.1, 0.2, 0.4, 0.6, 0.8))
}
\arguments{
\item{expr_file}{the address of the gene expressing file from EPIC dataset}

\item{expr_size}{expression bar width}

\item{expr_alpha}{expression bar alpha}

\item{mc.cores}{the CPU cores to use in the function}

\item{color_gradientn}{color gradients, a vector.}

\item{values}{controlling the distribution of colors, same length with color_gradientn}
}
\value{
local tree alignment result and corresponding gene expression. Phenotype and Genotype.
}
\description{
Add EPIC data to local tree alignment result
}
