% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvita-classic.R
\name{ggvita.classic}
\alias{ggvita.classic}
\title{Show alignment results in classical way: including expression and prune nodes}
\usage{
ggvita.classic(alml_list, result.order, expr_file, size, tip_size,
  values = c(0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8)/0.8,
  colors = rainbow(11)[1:9], mc.cores = 1, ...)
}
\arguments{
\item{alml_list}{the variable created from readal}

\item{result.order}{the order of result you want to show}

\item{size}{size of prune nodes and expression band width}

\item{tip_size}{size of tips}

\item{values}{the values for data distrbution adjustment.See scale_gradientn's values requirement}

\item{colors}{color gradinets. See scale_gradientn's color requirement}

\item{mc.cores}{This function is very slow. Suggest to use a few cores.}

\item{...}{same as ggtree parameters (ggtree::ggtree())}
}
\value{
a basic tree alignment structure
}
\description{
Show alignment results in classical way: including expression and prune nodes
}
