% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvita.R
\name{ggvita}
\alias{ggvita}
\title{ggvita}
\usage{
ggvita(one_result, branch_size = 0.2, branch_alpha = 1, tip_size = 0.5,
  tiplab_size = 0.1, print = F, mapping = NULL, layout = "rectangular",
  open.angle = 0, mrsd = NULL, as.Date = FALSE, yscale = "none",
  yscale_mapping = NULL, ladderize = TRUE, right = FALSE,
  branch.length = "branch.length", ...)
}
\arguments{
\item{one_result}{one in read_alml(...)$`result_list`.}

\item{tip_size}{tip size, default is 0.5.}

\item{tiplab_size}{tip label size, default is 0.1.}

\item{print}{Logical. If TRUE, then the alml result is printed. Otherwise, not.}

\item{layout}{one of 'rectangular', 'slanted', 'fan', 'circular', 'radial', 'equal_angle' or 'daylight'.}

\item{...}{same as the function ggtree attributes. See it by help(ggtree).}

\item{brach_size}{branch size, default is 0.2.}
}
\description{
Draw trees alignment results
}
