% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{ggvita_add}
\alias{ggvita_add}
\title{Add layer to ggvita}
\usage{
ggvita_add(e1, e2)
}
\arguments{
\item{e1}{element one, a ggvita object}

\item{e2}{element two, any layer in ggvita, ggtree, ggplot, including "geom_?", "stat_?", "scale_?","theme_?"}
}
\value{
a ggvita object
}
\description{
The key function which helps ggvita work together with ggtree and ggplot2 system. OO system. Layer adding style.
}
