% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LN_to_Bin.R
\name{LN_to_Bin}
\alias{LN_to_Bin}
\title{Transform Lineage Name to Lineage}
\usage{
LN_to_Bin(the_LN_vec, mc.cores = 1, Root.style = "Root")
}
\arguments{
\item{the_LN_vec}{C.elegans cell Lineage Name (e.g. "MSapapap")}

\item{mc.cores}{This function is very slow. Suggest to use a few cores.}
}
\value{
Lineage(binary sequence, index)
}
\description{
Transform C.elegans cell Lineage Name (e.g. "MSapapap") to Lineage ("10010101") (binary sequence, index)
}
\author{
Meng Yuan
}
