% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make-tree-complete.R
\name{Make.tree.complete}
\alias{Make.tree.complete}
\title{Make an incompleted tree completed}
\usage{
Make.tree.complete(allLin)
}
\arguments{
\item{allLin}{incompleted all lineages.}
}
\value{
completed all lineages.
}
\description{
Make an incompleted tree completed
}
