% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{+.ggvita}
\alias{+.ggvita}
\title{Add layer to ggvita}
\usage{
\method{+}{ggvita}(e1, e2)
}
\arguments{
\item{e1}{element one, a ggvita object}

\item{e2}{element two, any layer in ggvita, ggtree, ggplot, including "geom_*", "stat_*", "scale_*",....but not "theme_*" ("%+ggvita%" instead)}
}
\value{
a ggvita object
}
\description{
The key function which helps ggvita work together with ggtree and ggplot2 system. OO system. Layer adding style.
}
