% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readal.R
\name{readal.alm}
\alias{readal.alm}
\title{Read in an alm file}
\usage{
readal.alm(file)
}
\arguments{
\item{file}{the address of the alm file}
}
\value{
NULL, but create a variable named "alm_label" automately
}
\description{
Read in an alm file and create a list accordingly
}
\details{
An alm file contains the information of teriminal tips and their tissue types.
}
\keyword{readal.alm}
