% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readal.R
\name{readal.alml}
\alias{readal.alml}
\title{read_alml}
\usage{
readal.alml(file)
}
\arguments{
\item{file}{file address}
}
\value{
a list containing the result (phylos of treeS and treeT, the matching dataframe) and the result analysis (pValue and etc.)
}
\description{
Read local alignment results from HSA calculation and transform the result from txt to list in R
}
\author{
Meng Yuan
}
