% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readal-epic.R
\name{readal.epic}
\alias{readal.epic}
\title{Read in a gene expression level file on the tree downloaded from EPIC}
\usage{
readal.epic(expr_file, mc.cores = 1)
}
\arguments{
\item{expr_file}{a file downloaded from EPIC dataset, usually ending with .csv}

\item{mc.cores}{This function is very slow. Suggest to use a few cores.}
}
\value{
a data.frame contains 4 columns: cell,time,blot,Lineage
}
\description{
Read in a gene expression level file on the tree downloaded from EPIC
}
