% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readal.R
\name{readal}
\alias{readal}
\title{Read in the result of DELTA for analysis
DELTA parameter}
\usage{
readal(outfile, fileS, fileT, cost = NULL, method = NULL,
  max.target = NULL, test = NULL, all = NULL, prune = NULL)
}
\arguments{
\item{outfile}{<Path of DELTA result>}

\item{fileS}{<TreeS file path>}

\item{fileT}{<TreeT file path>}

\item{cost}{<Cost file path>}

\item{method}{<l or g>}

\item{test}{testNum}

\item{all}{<T or F>}

\item{prune}{pruneScore}

\item{max_target}{<target num for l>}
}
\description{
Read in the result of DELTA for analysis
DELTA parameter
}
