% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReLin.R
\name{ReLin}
\alias{ReLin}
\title{Renew the incomplete tree leave lineages}
\usage{
ReLin(LeafLin, UseSubRoot = T)
}
\arguments{
\item{LeafLin}{incomplete leave lineages}
}
\value{
a list inclued both new and past lineages data.frames for all and leaves
}
\description{
Renew the incomplete tree leave lineages
}
