% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_Bin_by_Dec}
\alias{sort_Bin_by_Dec}
\title{Sort the vector of binary sequences increasingly according to their decimal values}
\usage{
sort_Bin_by_Dec(seq_vector)
}
\arguments{
\item{seq_vector}{a vector containing some binary sequences}
}
\value{
the vector of binary sequences in increasing order
}
\description{
Sort the vector of binary sequences increasingly according to their decimal values
}
