% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_prune.R
\name{stat_prune}
\alias{stat_prune}
\title{stat_prune}
\usage{
stat_prune(ggvita.object, color = "red", size = 2)
}
\arguments{
\item{ggvita.object}{the ggvita object you want to add to}

\item{color}{color of prune node information}

\item{size}{size of prune node information}

\item{...}{same with geom_segment parameters}
}
\description{
Add prune node informantion to tree alignment result
}
