% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-phylo.R
\name{tr2phylo}
\alias{tr2phylo}
\title{Transform a tree seq split with space to a phylo}
\usage{
tr2phylo(seq, the_root)
}
\arguments{
\item{seq}{the lineage name  seq of all nodes in a tree, split with space}

\item{the_root}{the root of the tree}
}
\description{
Transform a tree seq split with space to a phylo
}
