% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-alm.R
\name{write.alm}
\alias{write.alm}
\title{Write alm data.frame to file.}
\usage{
write.alm(dt, file.name)
}
\arguments{
\item{dt}{alm data frame}

\item{file.name}{the address of alm file to be created.}
}
\description{
Write alm data.frame to file.
}
