% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvita.R
\name{ggvita}
\alias{ggvita}
\title{Show the basic structure of alignment results}
\usage{
ggvita(alml_list, result.order, ...)
}
\arguments{
\item{alml_list}{the variable created from readal}

\item{result.order}{the order of result you want to show}

\item{...}{same as ggtree parameters (ggtree::ggtree())}
}
\value{
a basic tree alignment structure
}
\description{
Show the basic structure of alignment results
}
