% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvita.R
\name{ggvita}
\alias{ggvita}
\title{Show the basic structure of alignment results}
\usage{
ggvita(alml_list, result.order, trace_down_for_pruned = T, ...)
}
\arguments{
\item{alml_list}{the object read from readal}

\item{result.order}{the order of result you want to show}

\item{trace_down_for_pruned}{default is True. if FALSE, ggvita will not catch the sister terminal cell type for mother of pruned terminal cells.}

\item{...}{same as ggtree parameters (ggtree::ggtree()) ladderize=T}
}
\value{
a basic tree alignment structure
}
\description{
Show the basic structure of alignment results
}
