% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readal.R
\name{readal}
\alias{readal}
\title{Read in the result of DELTA for analysis
DELTA parameter}
\usage{
readal(outfile, fileS, fileT, cost, method, max.target = NULL, test = NULL,
  all = NULL, prune = NULL)
}
\arguments{
\item{outfile}{<Path of DELTA result>. Required.}

\item{fileS}{<TreeS file path>. Required.}

\item{fileT}{<TreeT file path>. Required.}

\item{cost}{<Cost file path>. Required.}

\item{method}{<l or g>}

\item{test}{testNum}

\item{all}{<T or F>}

\item{prune}{pruneScore}

\item{max_target}{<target num for l>}
}
\description{
Read in the result of DELTA for analysis
DELTA parameter
}
